4.21*: Blog List Expansion, step9

+++ part4/bloglist/controllers/blogs.js
@@ -40,6 +40,16 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user.id) {
+    return response.status(401).json({ error: 'user invalid' })
+  }
+
   try {
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
